rename public listLists to listLiveLists
